+++ taskfile/merge.go
@@ -9,7 +9,7 @@ import (
 const NamespaceSeparator = ":"
 
 // Merge merges the second Taskfile into the first
-func Merge(t1, t2 *Taskfile, internal bool, namespaces ...string) error {
+func Merge(t1, t2 *Taskfile, includedTaskfile *IncludedTaskfile, namespaces ...string) error {
 	if t1.Version != t2.Version {
 		return fmt.Errorf(`task: Taskfiles versions should match. First is "%s" but second is "%s"`, t1.Version, t2.Version)
 	}
@@ -39,22 +39,38 @@ func Merge(t1, t2 *Taskfile, internal bool, namespaces ...string) error {
 		t1.Tasks = make(Tasks)
 	}
 	for k, v := range t2.Tasks {
-		// FIXME(@andreynering): Refactor this block, otherwise we can
-		// have serious side-effects in the future, since we're editing
-		// the original references instead of deep copying them.
+		// We do a deep copy of the task struct here to ensure that no data can
+		// be changed elsewhere once the taskfile is merged.
+		task := v.DeepCopy()
 
-		v.Internal = v.Internal || internal
+		// Set the task to internal if EITHER the included task or the included
+		// taskfile are marked as internal
+		task.Internal = task.Internal || includedTaskfile.Internal
 
-		t1.Tasks[taskNameWithNamespace(k, namespaces...)] = v
-
-		for _, dep := range v.Deps {
+		// Add namespaces to dependencies, commands and aliases
+		for _, dep := range task.Deps {
 			dep.Task = taskNameWithNamespace(dep.Task, namespaces...)
 		}
-		for _, cmd := range v.Cmds {
+		for _, cmd := range task.Cmds {
 			if cmd != nil && cmd.Task != "" {
 				cmd.Task = taskNameWithNamespace(cmd.Task, namespaces...)
 			}
 		}
+		for i, alias := range task.Aliases {
+			task.Aliases[i] = taskNameWithNamespace(alias, namespaces...)
+		}
+		// Add namespace aliases
+		if includedTaskfile != nil {
+			for _, namespaceAlias := range includedTaskfile.Aliases {
+				task.Aliases = append(task.Aliases, taskNameWithNamespace(task.Task, namespaceAlias))
+				for _, alias := range v.Aliases {
+					task.Aliases = append(task.Aliases, taskNameWithNamespace(alias, namespaceAlias))
+				}
+			}
+		}
+
+		// Add the task to the merged taskfile
+		t1.Tasks[taskNameWithNamespace(k, namespaces...)] = task
 	}
 
 	return nil