+++ taskfile/task.go
@@ -11,6 +11,7 @@ type Task struct {
 	Label                string
 	Desc                 string
 	Summary              string
+	Aliases              []string
 	Sources              []string
 	Generates            []string
 	Status               []string
@@ -56,6 +57,7 @@ func (t *Task) UnmarshalYAML(unmarshal func(interface{}) error) error {
 		Label         string
 		Desc          string
 		Summary       string
+		Aliases       []string
 		Sources       []string
 		Generates     []string
 		Status        []string
@@ -78,6 +80,7 @@ func (t *Task) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	t.Deps = task.Deps
 	t.Label = task.Label
 	t.Desc = task.Desc
+	t.Aliases = task.Aliases
 	t.Summary = task.Summary
 	t.Sources = task.Sources
 	t.Generates = task.Generates
@@ -94,4 +97,36 @@ func (t *Task) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	t.IgnoreError = task.IgnoreError
 	t.Run = task.Run
 	return nil
+}
+
+// DeepCopy creates a new instance of Task and copies
+// data by value from the source struct.
+func (t *Task) DeepCopy() *Task {
+	c := &Task{
+		Task:                 t.Task,
+		Cmds:                 deepCopySlice(t.Cmds),
+		Deps:                 deepCopySlice(t.Deps),
+		Label:                t.Label,
+		Desc:                 t.Desc,
+		Summary:              t.Summary,
+		Aliases:              deepCopySlice(t.Aliases),
+		Sources:              deepCopySlice(t.Sources),
+		Generates:            deepCopySlice(t.Generates),
+		Status:               deepCopySlice(t.Status),
+		Preconditions:        deepCopySlice(t.Preconditions),
+		Dir:                  t.Dir,
+		Vars:                 t.Vars.DeepCopy(),
+		Env:                  t.Env.DeepCopy(),
+		Silent:               t.Silent,
+		Interactive:          t.Interactive,
+		Internal:             t.Internal,
+		Method:               t.Method,
+		Prefix:               t.Prefix,
+		IgnoreError:          t.IgnoreError,
+		Run:                  t.Run,
+		IncludeVars:          t.IncludeVars.DeepCopy(),
+		IncludedTaskfileVars: t.IncludedTaskfileVars.DeepCopy(),
+		IncludedTaskfile:     t.IncludedTaskfile.DeepCopy(),
+	}
+	return c
 }